Remove `livePreview` and its implementation
fixes https://github.com/microsoft/vscode/issues/205535
